Proposal of Algorithm class (#179)
* initial revision

* Removed class members of Algorithm class

added update_objective

* initial version. Fix inline __idiv__

* First implementation of CompositeOperator/DataContainer

* removed __getitem__ added get_item added shape

* added CGLS

* working unit test, initial tomography test

* added reverse multiplication of operator with number

* added operators directory

* fixed typo

* added unittest for CompositeDataContainer

* fix TomoIdentity with scalar

* check numerical types from numpy

* add default stop criterion and run method

* add run method

* first working implementation of CGLS with CompositeOperator/DataContainer

notice problem with _rmul_ and _mul_ methods precedence with numpy.

* new Algorithm class and algorithms in separate files

Added new Algorithm class and derivatives in different files for
GradientDescent, CGLS, FBPD, FISTA

* added algorithms and restored CIL_VERSION env variable

* removed Algorithms.py

* modified run and renamed a few members/methods

* uses squared_norm

* renamed get_current_objective to get_last_objective

update_objective can be issued every N iteration, default 1.
fixed run method to run N iterations within the stop criterion.

* load class as module files

* force py line endings to LF

* updates

* call super __init__ as first thing

* unit tests are now to be found in test directory

unit tests are now split in several files in the directory test

* install algorithms module

* Implementation with Algorithm

* skip Reader tests

* unittest for linux

* commented not needed import Iterable

* removed explicit return from __init__

* remove composite operator file
