+++ .gitattributes
@@ -1,0 +1,1 @@
+*.py text eol=lf
