+++ Wrappers/Python/conda-recipe/meta.yaml
@@ -12,6 +12,15 @@ test:
   requires:
     - python-wget
     - cvxpy # [not win]
+
+  source_files:
+    - ./test # [win]
+    - ./ccpi/Wrappers/Python/test # [not win] 
+
+  commands:
+    - python -c "import os; print (os.getcwd())"
+    - python -m unittest discover # [win]
+    - python -m unittest discover -s ccpi/Wrappers/Python/test # [not win]
     
 requirements:
   build:
